service/dap: add support for basic log points
This uses the resume execution logic from PR #2530
and uses it to resume execution when a log point is
hit.

The debugger should not "break" on a log point, but
should instead log the user supplied message and then
resume execution. This change will do this by setting
Tracepoint on breakpoints with a log message. It will
not resume execution if a step was interrupted and
NextInProgress is false.

This PR does not attempt to handle evaluate expressions
within the log message, which can be specified with {}.

Updates golang/vscode-go#123